requirements.txt
Django==1.8
argparse==1.2.1
django-debug-toolbar==1.4
sqlparse==0.1.18
wsgiref==0.1.2
